+++ dependencies/engine/resources/source/common/ShaderBase.cpp
@@ -55,7 +55,7 @@ GLuint ShaderBase::loadShader(GLenum shaderType, const std::string &shaderSource
     GLuint shader = glCreateShader(shaderType);
     if (shader) {
         auto *shaderRawString = static_cast<const GLchar*>(shaderSource.c_str());
-        GLint shaderLength = shaderSource.length();
+        GLint shaderLength = static_cast<GLint>(shaderSource.length());
         glShaderSource(shader, 1, &shaderRawString, &shaderLength);
         glCompileShader(shader);
